Add cell dtype

+++ src/pygeogrids/netcdf.py
@@ -172,7 +172,7 @@ def save_lonlat(filename, arrlon, arrlat, geodatum, arrcell=None,
         setattr(longitude, 'valid_range', [-180.0, 180.0])
 
         if arrcell is not None:
-            cell = ncfile.createVariable('cell', np.dtype('int16').char,
+            cell = ncfile.createVariable('cell', np.dtype('int32').char,
                                          dim,
                                          shuffle=shuffle,
                                          zlib=zlib, complevel=complevel)
